cmp-carmackize: Look at correct output data when compressing
Was looking at the output (compressed) data, instead of the yet-to-be-processed
input data, at best providing suboptimal compression and at worst corrupting
data.

Also the RecordBuffer reads were replaced with Uint16Array types which are not
endian-safe but in this case doesn't matter and should increase the speed.

Debugging lines are also included for future debugging, but they are commented
out to avoid slowing down the browser when debugging unrelated issues.