cmp-rle-id: Correctly implement segment/offset normalization bug
Turns out it's not an extra byte inserted every 0xFF00 bytes, but rather if an
RLE code completes beyond every block of 0xFF00 bytes, the data is effectively
truncated to a multiple of 16 bytes due to a bug in the way the games adjust
the segment.  The next RLE code following this then continues on from the
truncation point, effectively overwriting some of the previous data.