Fix GIF LZW end-of-codestream 0-index bug
When using GIF-style LZW settings (and possibly with other settings
as well), a compressed data codestream that has a pending codeword
index of 0 at its end was not being encoded correctly. The simplest
example of this is attempting to encode the single byte 00 with these
settings.

The cause of this bug was an incomplete check of the idxPending
variable after the input stream ends; it should be checking for
non-null rather than non-zero.